refactor: delete widgets directory
