+++ src/components/map-widget/map-widget.tsx
@@ -1,9 +1,9 @@
 import 'leaflet/dist/leaflet.css';
 import {FC} from 'react';
 import {MapContainer, Marker, Popup, TileLayer} from 'react-leaflet';
-import {Coordinates} from '../../../shared/types/coordinates.ts';
-import {activeMarker, defaultMarker} from '../model/markers.ts';
-import {PointOnMap} from '../model/types.ts';
+import {Coordinates} from '../../shared/types/coordinates.ts';
+import {activeMarker, defaultMarker} from './model/markers.ts';
+import {PointOnMap} from './model/types.ts';
 import {UpdateMapCenter} from './update-map-center.tsx';
 
 const ATTRIBUTION = '&copy; <a href="https://www.openstreetmap.org/copyright">OpenStreetMap</a> contributors &copy; <a href="https://carto.com/attributions">CARTO</a>';
