+++ src/components/map-widget/model/types.ts
