+++ src/components/selector-widget/model/types.ts
