+++ src/pages/offer-page/comments-list.tsx
@@ -3,7 +3,7 @@ import {useParams} from 'react-router-dom';
 import {CommentaryCard} from '../../components/commentary-card/commentary-card.tsx';
 import {loadComments} from '../../slices/offer-page-slice.ts';
 import {useAppDispatch, useAppSelector} from '../../shared/redux-helpers/typed-hooks.ts';
-import {Spinner} from '../../widgets/spinner/ui/spinner.tsx';
+import {Spinner} from '../../components/spinner/spinner.tsx';
 
 export const CommentsList = () => {
   const dispatch = useAppDispatch();
