+++ src/pages/offer-page/offer-page.tsx
@@ -7,9 +7,9 @@ import {loadNearbyOffers, loadOffer} from '../../slices/offer-page-slice.ts';
 import {RoutePath} from '../../shared/enums/routes.ts';
 import {useAppDispatch, useAppSelector} from '../../shared/redux-helpers/typed-hooks.ts';
 import {FavoriteStatus} from '../../shared/server-interaction/constants.ts';
-import {PointOnMap} from '../../widgets/map/model/types.ts';
-import {MapWidget} from '../../widgets/map/ui/map-widget.tsx';
-import {Spinner} from '../../widgets/spinner/ui/spinner.tsx';
+import {PointOnMap} from '../../components/map-widget/model/types.ts';
+import {MapWidget} from '../../components/map-widget/map-widget.tsx';
+import {Spinner} from '../../components/spinner/spinner.tsx';
 import {CommentsList} from './comments-list.tsx';
 import {FeedbackForm} from './feedback-form.tsx';
 import {OfferPageWrapper} from './offer-page-wrapper.tsx';
